flake8; moved to test_models.py
